Move runtime support to guest staging
